Activate favourite screen
